Modifications and updates
Modified ForEachMessage to accept lowID, highID and suffix args and
removed the ForEachMessageBySuffix function as it was redundant. Added
GetLastID and GetLastIDBySuffix as they were needed by the IMAP server.
Also added versioning support for the data store. Updated .travis.yml
to clarify that we support only Go version >= 1.4. This is because of
the minimum Go version requirement of gRPC.